# openapi.yml
openapi: 3.0.3
info:
  title: Classroom API
  description: Google Classroom clone
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /courses/{courseId}/posts/{postId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"

    get:
      tags: [Posts]
      summary: Get post details
      operationId: getPost
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Posts]
      summary: Update a post (teacher only)
      operationId: updatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Posts]
      summary: Delete a post (teacher only)
      operationId: deletePost
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/posts/{postId}/materials:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"

    get:
      tags: [Post Materials]
      summary: List materials attached to a post
      operationId: listPostMaterials
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Post Materials]
      summary: Upload a material to a post (teacher only)
      operationId: uploadPostMaterial
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
      responses:
        "201":
          description: Material uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{courseId}/posts/{postId}/materials/{fileId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/FileId"

    delete:
      tags: [Post Materials]
      summary: Remove a material (teacher only)
      operationId: deletePostMaterial
      responses:
        "204":
          description: Material deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/posts/{postId}/materials/{fileId}/download:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/FileId"

    get:
      tags: [Post Materials]
      summary: Download a material file
      operationId: downloadPostMaterial
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /courses/{courseId}/posts/{postId}/solutions:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"

    get:
      tags: [Solutions]
      summary: List all solutions for a task (teacher only)
      operationId: listSolutions
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SolutionStatus"
          description: Filter by solution status
      responses:
        "200":
          description: Paginated solutions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageSolutionDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Solutions]
      summary: Submit a solution for a task (student only)
      operationId: createSolution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSolutionRequest"
      responses:
        "201":
          description: Solution submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionDto"
        "400":
          description: Post is not a task or deadline passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Solution already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{courseId}/posts/{postId}/solutions/my:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"

    get:
      tags: [Solutions]
      summary: Get the current student's solution for a task
      operationId: getMySolution
      responses:
        "200":
          description: Student's solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/posts/{postId}/solutions/{solutionId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/SolutionId"

    get:
      tags: [Solutions]
      summary: Get a specific solution (teacher or owner)
      operationId: getSolution
      responses:
        "200":
          description: Solution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Solutions]
      summary: Update solution text (student, own only)
      operationId: updateSolution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSolutionRequest"
      responses:
        "200":
          description: Solution updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Solutions]
      summary: Delete a solution (student, own only)
      operationId: deleteSolution
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/posts/{postId}/solutions/{solutionId}/files:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/SolutionId"

    get:
      tags: [Solution Files]
      summary: List files attached to a solution
      operationId: listSolutionFiles
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Solution Files]
      summary: Upload a file to a solution (student, own only)
      operationId: uploadSolutionFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDto"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses/{courseId}/posts/{postId}/solutions/{solutionId}/files/{fileId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/SolutionId"
      - $ref: "#/components/parameters/FileId"

    delete:
      tags: [Solution Files]
      summary: Remove a file from a solution (student, own only)
      operationId: deleteSolutionFile
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/posts/{postId}/solutions/{solutionId}/files/{fileId}/download:
    parameters:
      - $ref: "#/components/parameters/CourseId"
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/SolutionId"
      - $ref: "#/components/parameters/FileId"

    get:
      tags: [Solution Files]
      summary: Download a solution file
      operationId: downloadSolutionFile
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SolutionId:
      name: solutionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PostType:
      type: string
      enum: [MATERIAL, TASK]
    SolutionStatus:
      type: string
      enum: [SUBMITTED, GRADED]
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: johndoe
        displayName:
          type: string
          example: John Doe
        createdAt:
          type: string
          format: date-time
...
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        content:
          type: string
          maxLength: 10000
        deadline:
          type: string
          format: date-time
          nullable: true
    PostDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Week 3 — Sorting Algorithms"
        content:
          type: string
        type:
          $ref: "#/components/schemas/PostType"
        deadline:
          type: string
          format: date-time
          nullable: true
        author:
          $ref: "#/components/schemas/UserDto"
        materialsCount:
          type: integer
          description: Number of attached files
        commentsCount:
          type: integer
        solutionsCount:
          type: integer
          description: Only present for TASK posts (teacher view)
        mySolutionId:
          type: string
          format: uuid
          nullable: true
          description: Current student's solution id, null if not submitted
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
          example: lecture_03.pdf
        contentType:
          type: string
          example: application/pdf
        sizeBytes:
          type: integer
          format: int64
          example: 204800
        uploadedAt:
          type: string
          format: date-time
...
    CreateSolutionRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 10000
          description: Solution text / answer body
          example: "My implementation uses merge sort because..."
    SolutionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        status:
          $ref: "#/components/schemas/SolutionStatus"
        grade:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: "Null until graded; 0-100"
        filesCount:
          type: integer
        student:
          $ref: "#/components/schemas/UserDto"
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        gradedAt:
          type: string
          format: date-time
          nullable: true
    PageSolutionDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SolutionDto"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Course not found
        path:
          type: string
          example: /api/v1/courses/999